+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add `GraphQLError`s to the response via `ResolutionEnvironment` [#346](https://github.com/leangen/graphql-spqr/issues/346)
+
 ## [0.10.1] - 2019-12-30
 ### Added
 - Easy hierarchical names for nested types: `DefaultTypeInfoGenerator#withHierarchicalNames` [#310](https://github.com/leangen/graphql-spqr/issues/310)